+++ frontend/src/components/profile/DeleteAccount.jsx
@@ -3,8 +3,9 @@ import axios from "axios";
 import Loader from "../common/Loader.jsx";
 import { showErrorToast, showSuccessToast } from "../../utils/ToastUtils.jsx";
 import { useNavigate } from "react-router-dom";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { authActions } from "../../store/authSlice.js";
+import Button from "../common/Button.jsx";
 
 const API_URL = import.meta.env.VITE_API_URL || "/api";
 
@@ -16,7 +17,20 @@ const DeleteAccount = () => {
 
   const [loading, setLoading] = useState(false);
 
+  const { isLoggedIn, isDemo } = useSelector((state) => state.auth);
+
+  const handleNavigate = () => {
+    if (!isDemo) {
+      navigate("/profile");
+    }
+  };
+
   const handleDeleteAccount = async () => {
+    if (isDemo) {
+      showErrorToast("Deleting account is not allowed in demo mode");
+      return;
+    }
+
     setLoading(true);
     try {
       const res = await axios.delete(`${API_URL}/api/v1/user/delete/account`, {
@@ -46,7 +60,7 @@ const DeleteAccount = () => {
         </div>
       )}
 
-      <div className="border border-red-200 rounded-xl p-5 sm:p-6 bg-red-50">
+      <div className="border border-red-300 rounded-xl p-5 sm:p-6 bg-red-50">
         {/* Header */}
         <h2 className="text-xl font-semibold text-red-600 mb-3">
           Delete Account
@@ -61,7 +75,7 @@ const DeleteAccount = () => {
         {/* Warning Box */}
         <div className="bg-red-100 border border-red-300 rounded-md p-4 mb-5">
           <p className="text-sm text-red-700 font-medium">
-            ⚠ Type <span className="font-bold">PASSWORD</span> to confirm.
+            ⚠ Enter your password to confirm.
           </p>
         </div>
 
@@ -70,21 +84,44 @@ const DeleteAccount = () => {
           placeholder="Enter your password to confirm"
           value={password}
           onChange={(e) => setPassword(e.target.value)}
-          className="w-full h-11 px-4 border rounded-lg mb-4 focus:outline-none "
+          disabled={isDemo}
+          className={`w-full h-11 px-4 border rounded-lg mb-4
+            ${
+              isDemo
+                ? "bg-gray-100 cursor-not-allowed text-gray-500"
+                : "focus:outline-none"
+            }`}
         />
 
         {/* Buttons */}
         <div className="flex flex-col sm:flex-row gap-3">
-          <button className="w-full sm:flex-1 h-11 bg-gray-200 hover:bg-gray-300 text-gray-800 font-medium rounded-lg transition">
-            Cancel
-          </button>
+          {isLoggedIn && (
+            <Button
+              onClick={handleNavigate}
+              className={`w-full sm:flex-1 h-11 font-medium rounded-lg transition
+                 ${
+                   isDemo
+                     ? "bg-gray-400 cursor-not-allowed"
+                     : "bg-gray-200 hover:bg-gray-300 text-gray-800"
+                 }`}
+            >
+              Cancel
+            </Button>
+          )}
 
-          <button
-            className="w-full sm:flex-1 h-11 bg-red-600 hover:bg-red-700 text-white font-medium rounded-lg transition"
-            onClick={handleDeleteAccount}
-          >
-            Delete Account
-          </button>
+          {isLoggedIn && (
+            <Button
+              onClick={handleDeleteAccount}
+              className={`w-full sm:flex-1 h-11 font-medium rounded-lg transition
+                 ${
+                   isDemo
+                     ? "bg-gray-400 cursor-not-allowed"
+                     : "bg-red-600 hover:bg-red-700 text-white"
+                 }`}
+            >
+              Delete Account
+            </Button>
+          )}
         </div>
       </div>
     </>
